Better integral/derivative, change sample time type from int to double.

+++ PID/src/pl/krzysiekgl/Derivative.java
@@ -1,15 +1,15 @@
 package pl.krzysiekgl;
 
 public final class Derivative {
-    private int Ts;             //sample time in milliseconds
+    private double Ts;             //sample time in milliseconds
     private double prevValue;   //previous value necessary for calculations
     
-    public Derivative(int sampleTime, double initialCondition) {
+    public Derivative(double sampleTime, double initialCondition) {
         this.Ts = sampleTime;
         prevValue = initialCondition;
     }
     
-    public Derivative(int sampleTime) {
+    public Derivative(double sampleTime) {
         this(sampleTime, 0);
     }
     
@@ -21,7 +21,7 @@ public final class Derivative {
     }
     
     //Getters
-    public int getTs() {
+    public double getTs() {
         return Ts;
     }
